requirements.txt
pynput
Pillow






